Rename JukeboxSong and StaticJukeboxSong
